pr.machine field added
this version of the plate reader code was updated to contain the
property machine, which finds the serial number of the device and saves
it in pr.serialnumber, and tells you which PR it belongs to in
pr.machine.

added function ppf.transformEach which creates individual fits strain
by strain and provides a report on the variance of such fits.

ppf.robustCorrect is a little trick to avoid the LinAlgError that
arises while fitting, so just catches those errors and carries on.