working alignment to GR and plotReplicateMeanNew improvement
alignAll now functional.

now xpr.plotReplicateMeanNew() returns the data used for plotting,
whether it is the interpolation or the means of the means of every
bootstrap. It also automatically excludes replicates of the null
condition when aligning by growth rate.

alignAll used to exclude the wild type. now it doesn’t.